update inputs to new locations in vk83

+++ setup_cice_restarts.sh
@@ -2,10 +2,12 @@
 # patch for https://github.com/open-mpi/ompi/issues/12141
 if ! [ -f work/access-om3.cice.r.* ]
 then 
-# no restart files yet, use initial conditions
+# no restart files yet, use initial conditions if they exist
 IC=$(readlink work/input/iced.1900-01-01-10800.nc)
-rm work/input/iced.1900-01-01-10800.nc
-cp $IC work/input/iced.1900-01-01-10800.nc
+if [ -n "$IC" ]; then
+    rm work/input/iced.1900-01-01-10800.nc
+    cp $IC work/input/iced.1900-01-01-10800.nc
+fi
 else
 # change restart symlink to hardlink
 RESTART=$(echo work/access-om3.cice.r.*)
